requirements.txt
python-telegram-bot==20.7
feedparser
deep-translator
requests
beautifulsoup4